+++ Utils.h
@@ -8,4 +8,4 @@ float rndfMinMax(const float min, const float max);
 unsigned int rndMinMax(const unsigned int min, const unsigned int max);
 bool rndYesNo();
 
-float getTime();
+double getTime();
